adds synced and bootstrapping plugin to integration test networks
