+++ tools/integration-tests/tester/docker-compose.yml
@@ -7,7 +7,6 @@ services:
     working_dir: /go/src/github.com/iotaledger/goshimmer/tools/integration-tests/tester
     entrypoint: go test ./tests -v -mod=readonly
     volumes:
-    - /var/run/docker.sock:/var/run/docker.sock:ro
-    - ../../..:/go/src/github.com/iotaledger/goshimmer:ro
-    - ../logs:/tmp/logs
-
+      - /var/run/docker.sock:/var/run/docker.sock:ro
+      - ../../..:/go/src/github.com/iotaledger/goshimmer:ro
+      - ../logs:/tmp/logs
