+++ tools/integration-tests/tester/tests/synchronization_test.go
@@ -18,6 +18,9 @@ func TestNodeSynchronization(t *testing.T) {
 	require.NoError(t, err)
 	defer n.Shutdown()
 
+	// wait for peers to change their state to synchronized
+	time.Sleep(5 * time.Second)
+
 	numMessages := 100
 	idsMap := make(map[string]MessageSent, numMessages)
 	ids := make([]string, numMessages)
@@ -40,7 +43,7 @@ func TestNodeSynchronization(t *testing.T) {
 	checkForMessageIds(t, n.Peers(), ids, idsMap)
 
 	// spawn peer without knowledge of previous messages
-	newPeer, err := n.CreatePeer()
+	newPeer, err := n.CreatePeer(false)
 	require.NoError(t, err)
 	err = n.WaitForAutopeering(3)
 	require.NoError(t, err)
@@ -94,7 +97,7 @@ func checkForMessageIds(t *testing.T, peers []*framework.Peer, ids []string, ids
 		// check that the peer sees itself as synchronized
 		info, err := peer.Info()
 		require.NoError(t, err)
-		require.True(t, info.Synchronized)
+		require.True(t, info.Synced)
 
 		resp, err := peer.FindMessageByID(ids)
 		require.NoError(t, err)
